refactor(FE) : 파일 이름 위치 변경/GYMMATE-293
