+++ src/components/molecules/AddressInput.tsx
@@ -4,7 +4,7 @@ import React, { useState, ChangeEvent } from 'react';
 import { MagnifyingGlassIcon } from '@heroicons/react/24/outline';
 
 import InputField from '@/app/login/components/InputField';
-import AddressSearchModal from '@/components/molecules/AddressSearchModal/AddressSearchModal';
+import AddressSearchModal from '@/components/molecules/AddressSearchModal';
 
 interface AddressInputProps {
   setAddress: React.Dispatch<React.SetStateAction<string>>; // setAddress의 타입 정의
