+++ src/components/templates/ScheduleView/index.tsx
@@ -2,7 +2,7 @@
 
 import Calendar from '@/components/molecules/Calendar';
 import CalenderView from '@/components/templates/ScheduleView/CalenderView';
-import CalendarSummaryCards from '@/app/myfitness/components/CalendarCards/CalenderCards';
+import SummaryCard from '@/app/myfitness/components/CalenderCards';
 import ThreeLiftChartsTemplate from '@/components/molecules/Chart';
 
 const ScheduleView = () => {
@@ -14,10 +14,10 @@ const ScheduleView = () => {
       value: '2500KG',
       icon: (
         <img
-          src="/assets/icons/benchPress.svg"
           alt="벤치 프레스"
-          width={32}
           height={32}
+          src="/assets/icons/benchPress.svg"
+          width={32}
         />
       ),
     },
@@ -26,10 +26,10 @@ const ScheduleView = () => {
       value: '2500KG',
       icon: (
         <img
-          src="/assets/icons/deadlift.svg"
           alt="데드리프트"
-          width={32}
           height={32}
+          src="/assets/icons/deadlift.svg"
+          width={32}
         />
       ),
     },
@@ -38,10 +38,10 @@ const ScheduleView = () => {
       value: '2500KG',
       icon: (
         <img
-          src="/assets/icons/squat.svg"
           alt="스쿼트"
-          width={32}
           height={32}
+          src="/assets/icons/squat.svg"
+          width={32}
         />
       ),
     },
@@ -57,7 +57,7 @@ const ScheduleView = () => {
         <CalenderView />
       </div>
 
-      <CalendarSummaryCards data={summaryData} />
+      <SummaryCard data={summaryData} />
 
       <ThreeLiftChartsTemplate />
     </div>
